[Etapa 04] Ajustando função de inserção de declarações na lista de declarações global. Ajustando função de imprimir lista.

+++ include/comp_list.h
@@ -12,7 +12,7 @@ typedef struct _list
 {
 	int tipoVar;
 	char *nomeVar;
-	float tipoGlobal;
+   int tipoGlobal;
 	struct _list *next;
 } comp_list_t;
